pyproject.toml
[project]
name = "social-network-api"
version = "0.1.0"
description = "Production-ready social network API with FastAPI, Kafka, and PostgreSQL"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "SQLAlchemy[asyncio]==2.0.36",
    "asyncpg==0.29.0",
    "alembic==1.14.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "aiokafka==0.11.0",
    "redis==5.2.0",
    "structlog==24.4.0",
    "prometheus-fastapi-instrumentator==6.1.0",
    "opentelemetry-sdk==1.27.0",
    "opentelemetry-exporter-otlp==1.27.0",
    "opentelemetry-instrumentation-fastapi==0.48b0",
    "opentelemetry-instrumentation-asyncpg==0.48b0",
    "opentelemetry-instrumentation-redis==0.48b0",
    "opentelemetry-instrumentation-aio-pika==0.48b0",
    "httpx==0.27.2",
    "python-multipart==0.0.12",
    "jinja2==3.1.4",
    "email-validator==2.2.0",
    "orjson==3.10.7",
    "tenacity==9.0.0",
    "types-redis==4.6.0.20241004",
    "aiosqlite==0.20.0",
    "bcrypt==4.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "pytest-httpx==0.32.0",
    "testcontainers[postgres,kafka,redis]==4.9.1",
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "mypy==1.13.0",
    "types-passlib==1.7.7.20240819",
    "types-python-jose==3.3.4.20240106",
    "coverage==7.6.4",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --color=yes"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
markers = [
    "integration: mark tests that hit integration paths",
]

[tool.coverage.run]
branch = true
source = ["src/app"]
omit = [
    "tests/*",
    "src/app/main.py",
    "src/app/observability/*",
    "src/app/events/*",
    "src/app/db/*",
]

[tool.coverage.report]
fail_under = 85

[tool.coverage.html]
directory = "htmlcov"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
strict_optional = true
show_error_codes = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
